+++ src/config/config.cc
@@ -217,6 +217,7 @@ void inheritable_diagram_options::inherit(
         parent.generate_condition_statements);
     debug_mode.override(parent.debug_mode);
     generate_metadata.override(parent.generate_metadata);
+    type_aliases.override(parent.type_aliases);
 }
 
 std::string inheritable_diagram_options::simplify_template_type(