+++ src/config/option.h
@@ -25,6 +25,18 @@ namespace config {
 
 template <typename T> void append_value(T &l, const T &r) { l = r; }
 
+template <typename T>
+void append_value(std::vector<T> &l, const std::vector<T> &r)
+{
+    l.insert(std::end(l), r.begin(), r.end());
+}
+
+template <typename K, typename V>
+void append_value(std::map<K, V> &l, const std::map<K, V> &r)
+{
+    l.insert(r.begin(), r.end());
+}
+
 /**
  * Possible option inheritance methods from top level to diagram level.
  */
@@ -102,9 +114,7 @@ template <typename T> struct option {
             has_value = true;
         }
         else if (!is_declared && o.is_declared) {
-            value = o.value;
-            is_declared = true;
-            has_value = true;
+            set(o.value);
         }
     }