+++ tests/test_cases.cc
@@ -458,6 +458,7 @@ using namespace clanguml::test::matchers;
 #include "t20036/test_case.h"
 #include "t20037/test_case.h"
 #include "t20038/test_case.h"
+#include "t20039/test_case.h"
 
 ///
 /// Package diagram tests