+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
- 
+
+ * Enabled type_aliases config option for sequence diagrams (#224)
  * Refactored and unified JSON generators output (#223)
  * Added support for C++20 module based packages in class diagrams (#101)
  * Added support for class diagram filtering based on C++20 modules (#195)